refactor(event-stream): extract replay buffer to replay.rs
Pure code motion for #6235. Move the replay-buffer state machine out of
mod.rs into a dedicated replay.rs submodule: push_replay_frame
(admission, with buffer-full eviction), evict_replay_frame (the only
telemetry-loss removal, counts frames_replay_evicted and #2875 pause
poison), pop_replay_frame (non-loss removal for ACK-trim / shutdown), and
release_replay_dataplane_event_queue_budget (shutdown bulk drain). Every
removal releases the frame's queue budget exactly once.

Moved verbatim (no logic change). The three cross-module functions are
widened from private to pub(super); evict_replay_frame stays private
since only push_replay_frame calls it. mod.rs re-imports the three so the
connection / control / drain callers and the replay_budget test suite
(via `use super::*`) resolve unchanged. cargo build green.
